2nd_Semester/final_code/requirement.txt
addict==2.4.0
albumentations==1.1.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
Django==3.2.9
einops==0.3.2
facenet-pytorch==2.5.2
fonttools==4.28.2
idna==3.3
imageio==2.11.1
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.0
mmcv==1.3.17
networkx==2.6.3
numpy==1.21.4
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
regex==2021.11.10
requests==2.26.0
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.2
setuptools-scm==6.3.2
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
typing_extensions==4.0.0
urllib3==1.26.7
yapf==0.31.0
